[AM][OPS-10184] move high level interface of legacy calculator to calculator service from controller
