+++ app/ssttpcalculator/legacy/CalculatorService.scala
@@ -24,12 +24,14 @@ import ssttpcalculator.{DurationService, InterestRateService}
 import ssttpcalculator.legacy.CalculatorService.asTaxLiability
 import ssttpcalculator.legacy.model.TaxPaymentPlan
 import ssttpcalculator.legacy.model.TaxLiability.{amortizedLiabilities, latePayments}
-import ssttpcalculator.model.{Debit, Instalment, PaymentSchedule}
+import ssttpcalculator.model.{Debit, Instalment, PaymentPlanOption, PaymentSchedule}
 import ssttpcalculator.legacy.model.{LatePayment, Payment, TaxLiability}
 import times.ClockProvider
 import timetopaytaxpayer.cor.model.SelfAssessmentDetails
 import timetopaytaxpayer.cor.model.{Debit => CorDebit}
 import uk.gov.hmrc.selfservicetimetopay.models.PaymentDayOfMonth
+import _root_.model.PaymentScheduleExt
+import ssttpcalculator.model.PaymentPlanOption.{Additional, Basic, Higher}
 
 import java.time.LocalDate.now
 import java.time.temporal.ChronoUnit.DAYS
@@ -56,6 +58,44 @@ class CalculatorService @Inject() (
   val defaultInitialPaymentDays: Int = 11
   val `14 day gap between the initial payment date and the first scheduled payment date`: Int = 14
   val LastPaymentDelayDays = 7
+
+  def computeClosestSchedule(amount: BigDecimal, schedules: Seq[PaymentSchedule]): Option[PaymentSchedule] = {
+      def difference(schedule: PaymentSchedule) = math.abs(schedule.getMonthlyInstalment.toInt - amount.toInt)
+
+      def closest(min: PaymentSchedule, next: PaymentSchedule) = if (difference(next) < difference(min)) next else min
+
+      def lessThan(schedule: PaymentSchedule): Boolean = amount - schedule.instalmentAmount >= 0
+
+    schedules.filter(lessThan).reduceOption(closest)
+  }
+
+  def computeClosestSchedules(
+      maybeClosestSchedule: Option[PaymentSchedule],
+      schedules:            List[PaymentSchedule],
+      sa:                   SelfAssessmentDetails): Map[PaymentPlanOption, PaymentSchedule] = {
+    val scheduleList = maybeClosestSchedule match {
+      case None => List()
+      case Some(closestSchedule) => {
+        val closestScheduleIndex = schedules.indexOf(closestSchedule)
+
+          def scheduleMonthsShorter(n: Int): Option[PaymentSchedule] = closestScheduleIndex match {
+            case -1             => None
+            case i if i - n < 0 => None
+            case m              => Some(schedules(m - n))
+          }
+
+        if (closestScheduleIndex == 0) {
+          List(Some(closestSchedule))
+        } else if (closestScheduleIndex == 1)
+          List(Some(closestSchedule), scheduleMonthsShorter(1))
+        else
+          List(Some(closestSchedule), scheduleMonthsShorter(1), scheduleMonthsShorter(2))
+      }.flatten
+    }
+    val paymentPlanOptionReferences: Seq[PaymentPlanOption] = Seq(Basic, Higher, Additional)
+
+    paymentPlanOptionReferences.zip(scheduleList).toMap
+  }
 
   def computeSchedule(journey: Journey)(implicit request: Request[_]): PaymentSchedule = {
     val availableSchedules: Seq[PaymentSchedule] = availablePaymentSchedules(
