+++ app/views/calculator/legacy/how_much_can_you_pay_each_month_form_legacy.scala.html
@@ -67,7 +67,6 @@
     )
 })}
 
-
 @main(
     title = Messages("ssttp.calculator.results.title"),
     hasErrors = selectPlanForm.hasErrors
@@ -87,8 +86,6 @@
         )
     )
 }
-
-
 
 @viewsHelpers.form(redirectCall, 'id -> "paymentTodayForm") {
     <fieldset class="govuk-fieldset">
